refactor(#2456): split mujoco_viewer.py into backend/coordinator modules
- Extract VisualizationFlags, URDFToMJCFConverter, MuJoCoOffscreenRenderer
  to _mujoco_viewer_backend.py (533 LOC)
- Slim mujoco_viewer.py coordinator to 582 LOC (was 1071 LOC);
  retains MuJoCoViewerWidget + re-exports
- All public names re-exported for backward compatibility
- TDD contract tests added (7 tests, all GREEN)
